+++ src/task_monitor/mod.rs
@@ -3,16 +3,12 @@ use std::time::Duration;
 
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
-use once_cell::sync::Lazy;
-use prometheus::{linear_buckets, register_gauge, register_histogram, Gauge, Histogram};
 use tokio::select;
 use tokio::sync::{mpsc, watch, Mutex, Notify};
 use tokio::task::JoinHandle;
 use tracing::{error, info, instrument, warn};
 
 use crate::app::App;
-use crate::database::methods::DbMethods as _;
-use crate::database::Database;
 use crate::shutdown::Shutdown;
 
 pub mod tasks;
@@ -25,27 +21,6 @@ const FINALIZE_IDENTITIES_BACKOFF: Duration = Duration::from_secs(5);
 const QUEUE_MONITOR_BACKOFF: Duration = Duration::from_secs(5);
 const MODIFY_TREE_BACKOFF: Duration = Duration::from_secs(5);
 const SYNC_TREE_STATE_WITH_DB_BACKOFF: Duration = Duration::from_secs(5);
-
-static PENDING_IDENTITIES: Lazy<Gauge> = Lazy::new(|| {
-    register_gauge!("pending_identities", "Identities not submitted on-chain").unwrap()
-});
-
-static UNPROCESSED_IDENTITIES: Lazy<Gauge> = Lazy::new(|| {
-    register_gauge!(
-        "unprocessed_identities",
-        "Identities not processed by identity committer"
-    )
-    .unwrap()
-});
-
-static BATCH_SIZES: Lazy<Histogram> = Lazy::new(|| {
-    register_histogram!(
-        "submitted_batch_sizes",
-        "Submitted batch size",
-        linear_buckets(f64::from(1), f64::from(1), 100).unwrap()
-    )
-    .unwrap()
-});
 
 /// A task manager for all long running tasks
 ///
@@ -211,7 +186,7 @@ impl TaskMonitor {
              }
             // Or wait for a task to panic
             _ = Self::await_task_panic(&mut handles, shutdown.clone()) => {}
-        };
+        }
     }
 
     async fn await_task_panic(handles: &mut FuturesUnordered<JoinHandle<()>>, shutdown: Shutdown) {
@@ -233,28 +208,5 @@ impl TaskMonitor {
         if !shutdown.is_shutting_down() {
             warn!("all tasks have returned unexpectedly");
         }
-    }
-
-    async fn log_pending_identities_count(database: &Database) -> anyhow::Result<()> {
-        let identities = database.count_pending_identities().await?;
-        PENDING_IDENTITIES.set(f64::from(identities));
-        Ok(())
-    }
-
-    async fn log_unprocessed_identities_count(database: &Database) -> anyhow::Result<()> {
-        let identities = database.count_unprocessed_identities().await?;
-        UNPROCESSED_IDENTITIES.set(f64::from(identities));
-        Ok(())
-    }
-
-    async fn log_identities_queues(database: &Database) -> anyhow::Result<()> {
-        TaskMonitor::log_unprocessed_identities_count(database).await?;
-        TaskMonitor::log_pending_identities_count(database).await?;
-        Ok(())
-    }
-
-    #[allow(clippy::cast_precision_loss)]
-    fn log_batch_size(size: usize) {
-        BATCH_SIZES.observe(size as f64);
     }
 }
